+++ OpenSim/Simulation/MarkersReference.h
@@ -140,11 +140,12 @@ public:
         Therefore, subsequent changes to the Set of MarkerWeights will have
         no effect on the marker weights associated with this Reference. */
     void setMarkerWeightSet(const Set<MarkerWeight>& markerWeights);
-    void setDefaultWeight(double weight) { set_default_weight(weight); }
+    void setDefaultWeight(double weight);
 
 private:
     void constructProperties();
     void populateFromMarkerData(const MarkerData& markerData);
+    void updateInternalWeights() const;
 
 private:
     // Use a specialized data structure for holding the marker data
@@ -152,7 +153,7 @@ private:
     // marker names inside the marker data
     SimTK::Array_<std::string> _markerNames;
     // corresponding list of weights guaranteed to be in the same order as names above
-    SimTK::Array_<double> _weights;
+    mutable SimTK::Array_<double> _weights;
 
 //=============================================================================
 };  // END of class MarkersReference